fix/review-upgrade

+++ controllers/reviews.js
@@ -9,7 +9,7 @@ const getAll = async (req, res) => {
 
     const result = await Review.find({}, "-createdAt -updatedAt", { skip, limit });
       
-    res.json(result)
+    res.status(200).json(result)
 }
 
 // ---------------------- Get User Review ---------------------
@@ -18,10 +18,10 @@ const getAuthReview = async (req, res) => {
   const {page = 1, limit = 15 } = req.query;
   const skip = (page - 1) * limit;
   const { _id } = req.user;
- 
-  const result = await Review.find({ "owner._id": _id }, "-createdAt -updatedAt", { skip, limit });
-  
-  res.status(200).json(result)
+
+  const result = await Review.find({"owner._id": _id}, "-createdAt -updatedAt" , { skip, limit });
+
+  res.status(200).json(result);
 }
 
 // ------------------ Add Review --------------------------
@@ -41,19 +41,22 @@ const addReview = async (req, res) => {
 
 // ----------------- update Review ---------------------------------
 const updateReview = async (req, res) => {
-  const { name, email, avatarURL } = req.user;
-  const owner = {
-  name,
-  avatarURL,
-  email,
-  _id: req.user._id,
-    };
-
   const { reviewId: _id } = req.params;
+  
+  const currentUserId  = req.user._id;
+  const review = await Review.findById(_id); 
+      if (!review) {
+    throw HttpError(404);
+  }
+  const idByOwnerReview = review.owner._id.toString()
+ 
+  if (idByOwnerReview !== currentUserId.toString()) {
+    throw HttpError(401);
+  }
    
-  const result = await Review.findOneAndUpdate({_id}, {...req.body, owner}, {new: true});
+  const result = await Review.findOneAndUpdate({_id}, {...req.body}, {new: true});
   if(!result) {
-      throw HttpError(404, "Not found");
+      throw HttpError(404);
     }
   
   res.status(200).json(result);
@@ -61,12 +64,26 @@ const updateReview = async (req, res) => {
 
 //  ---------------- delete Review -----------------------------
 const deleteReview = async (req, res) => {
-    const {reviewId: _id } = req.params;
-    const result = await Review.findOneAndDelete({_id})
-    if(!result) {
-      throw HttpError(404, "Not found");
+  const { reviewId: _id } = req.params;
+
+  const currentUserId  = req.user._id;
+  const review = await Review.findById(_id); 
+
+    if (!review) {
+    throw HttpError(404);
+  }
+
+  const idByOwnerReview = review.owner._id.toString()
+
+  if (idByOwnerReview !== currentUserId.toString()) {
+    throw HttpError(401);
+  }
+  
+  const result = await Review.findOneAndDelete({ _id })
+  if (!result) {
+      throw HttpError(404);
     }
-    res.status(200).json({
+    res.json({
       message: "Delete success"
     })
   }
